requirements.txt
python-dotenv
ipykernel
pandas 
numpy 
pymongo
pymongo[srv]==3.11
certifi
scikit-learn
# dill
pyaml
mlflow

#-e .    # setup.py file gets triggred.